binding.gyp - reorganization
- put sources in binding.gyp
- split event.cc into multiple files
- move isReadyToSample to bindings.cc
- add Event::makeFromOboeMetadata()
