+++ src/bindings.h
@@ -9,65 +9,7 @@
 
 typedef int (*send_generic_span_t) (char*, uint16_t, oboe_span_params_t*);
 
-//
-// Metadata - work with oboe's oboe_metadata_t structure.
-//
-class Metadata : public Napi::ObjectWrap<Metadata> {
 
-public:
-  Metadata();
-  Metadata(const Napi::CallbackInfo& info);
-  ~Metadata();
-
-  // keep a copy of the metadata.
-  oboe_metadata_t metadata;
-
-  // these functions operate on the metadata
-  bool sampleFlagIsOn();
-  Napi::Value isValid(const Napi::CallbackInfo& info);
-  Napi::Value getSampleFlag(const Napi::CallbackInfo& info);
-  Napi::Value setSampleFlagTo(const Napi::CallbackInfo& info);
-  Napi::Value toString(const Napi::CallbackInfo& info);
-
-  // return a Metadata object
-  static Napi::Value fromString(const Napi::CallbackInfo& info);
-  static Napi::Value makeRandom(const Napi::CallbackInfo& info);
-
-  // instanceof equivalent for C++ functions
-  static bool isMetadata(const Napi::Object);
-
-  // extract metadata from Napi::Value which may be an instance
-  // of Metadata, an Event, a string, or Napi::External referring
-  // to oboe_metadata_t.
-  static bool getMetadata(Napi::Value, oboe_metadata_t*);
-
-  // work with existing metadata
-  static Napi::Value sampleFlagIsSet(const Napi::CallbackInfo& info);
-  static bool format(oboe_metadata_t* md, size_t len, char* buffer, uint flags);
-  const static int ff_header = 1;
-  const static int ff_task = 2;
-  const static int ff_op = 4;
-  const static int ff_flags = 8;
-  const static int ff_sample = 16;
-  const static int ff_separators = 32;
-  const static int ff_lowercase = 64;
-
-  // predefined formats
-  const static int fmtHuman = ff_header | ff_task | ff_op | ff_flags | ff_separators | ff_lowercase;
-  const static int fmtLog = ff_task | ff_sample | ff_separators;
-
-  // size needed to format is the size needed + 3 for delimiters to split the parts.
-  const static size_t fmtBufferSize = OBOE_MAX_METADATA_PACK_LEN + 3;
-
-  // C++ callable for creating an instance of Metadata
-  static Napi::Object NewInstance(Napi::Env, Napi::Value);
-
-private:
-  static Napi::FunctionReference constructor;
-
-public:
-  static Napi::Object Init(Napi::Env, Napi::Object);
-};
 
 
 //
@@ -130,7 +72,10 @@ public:
   // C++ instanceof equivalent
   static bool isEvent(Napi::Object);
 
-private:
+  // C++ method to create an unitialized, invalid oboe event.
+  static Napi::Object makeFromOboeMetadata(const Napi::Env env, oboe_metadata_t& omd);
+
+ private:
   static Napi::FunctionReference constructor;
 
 public:
@@ -160,6 +105,12 @@ namespace Sanitizer {
   Napi::Object Init(Napi::Env, Napi::Object);
 }
 
+//
+// Notifier provides asynchronous notifications from oboe
+//
+namespace Notifier {
+  Napi::Object Init(Napi::Env, Napi::Object);
+}
 //
 // Config provides the getVersionString function.
 //
